完成了删除商品前对商品有无关联货道进行判断

+++ dkd-manage/src/main/java/com/dkd/manage/service/impl/SkuServiceImpl.java
@@ -1,7 +1,11 @@
 package com.dkd.manage.service.impl;
 
 import java.util.List;
+
+import com.dkd.common.exception.ServiceException;
 import com.dkd.common.utils.DateUtils;
+import com.dkd.manage.domain.Channel;
+import com.dkd.manage.mapper.ChannelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import com.dkd.manage.mapper.SkuMapper;
@@ -19,6 +23,8 @@ public class SkuServiceImpl implements ISkuService
 {
     @Autowired
     private SkuMapper skuMapper;
+    @Autowired
+    private ChannelMapper channelMapper;
 
     /**
      * 查询商品管理
@@ -79,6 +85,13 @@ public class SkuServiceImpl implements ISkuService
     @Override
     public int deleteSkuBySkuIds(Long[] skuIds)
     {
+        //1.判断当前商品是否有关联货道 如果有，则不被允许删除
+        int count = channelMapper.countChannelBySkuIds(skuIds);
+        if(count > 0){
+            throw new ServiceException("当前商品有关联货道，不允许删除");
+        }
+
+        //2.删除商品
         return skuMapper.deleteSkuBySkuIds(skuIds);
     }
